move `run_json_block()` to `tests/` (#17269)
