+++ tests/tools/test_run_block.py
@@ -10,7 +10,7 @@ from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
 from chia.util.ints import uint32, uint64, uint128
-from tools.run_block import run_json_block
+from tests.util.run_block import run_json_block
 
 constants = dataclasses.replace(
     DEFAULT_CONSTANTS,
